django/requirements.txt
Django>=4.2
django-environ
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
gunicorn
requests
PyJWT[crypto]
git+https://github.com/liqpay/sdk-python#egg=liqpay-python
ecdsa
